Upgrade php/compose/buildx/helm
* match versions between debian and alpine
* this means debian helm upgraded from helm 2 to 3
* adding support for arm64 (tho no mutagen for it)

+++ Dockerfile
@@ -1,5 +1,9 @@
 # syntax=docker/dockerfile:1.4
-FROM php:7.4-cli-alpine AS build
+ARG BUILDX_VERSION=0.9.1
+ARG COMPOSE_VERSION=2.14.0
+ARG HELM_VERSION=3.10.2
+ARG KUBESEAL_VERSION=0.17.5
+FROM --platform=$BUILDPLATFORM php:8.0-cli-alpine AS build
 ARG WS_VERSION=0.2.x
 
 RUN apk add --no-cache bash git icu-dev
@@ -30,14 +34,22 @@ RUN <<EOF
   # workspace
   wget -O /tmp/ws.tar.gz "https://github.com/my127/workspace/archive/${WS_VERSION}.tar.gz"
   tar -C /usr/src -xvf /tmp/ws.tar.gz
-  cd "/usr/src/workspace-${WS_VERSION}"
+  mv "/usr/src/workspace-${WS_VERSION}" /usr/src/workspace
+  cd /usr/src/workspace
   export COMPOSER_ALLOW_SUPERUSER=1
   composer install
   composer compile
 EOF
 
-FROM php:7.4-cli-alpine as alpine
-ARG WS_VERSION=0.2.x
+FROM docker/buildx-bin:$BUILDX_VERSION as buildx
+
+FROM php:8.0-cli-alpine as alpine
+ARG TARGETARCH
+ARG COMPOSE_VERSION
+ARG HELM_VERSION
+ARG KUBESEAL_VERSION
+
+COPY --from=buildx /buildx /usr/libexec/docker/cli-plugins/docker-buildx
 
 RUN <<EOF
   set -o errexit
@@ -56,63 +68,79 @@ RUN <<EOF
 
   # docker compose v2
   mkdir -p /usr/libexec/docker/cli-plugins
-  wget -O /usr/libexec/docker/cli-plugins/docker-compose https://github.com/docker/compose/releases/download/v2.12.1/docker-compose-linux-x86_64
+  wget -O /usr/libexec/docker/cli-plugins/docker-compose "https://github.com/docker/compose/releases/download/v${COMPOSE_VERSION}/docker-compose-linux-$(uname -m)"
   chmod +x  /usr/libexec/docker/cli-plugins/docker-compose
 
   # helm2
-  wget -O helm.tar.gz "https://get.helm.sh/helm-v2.17.0-linux-amd64.tar.gz"
-  tar -C /usr/local/bin --strip-components=1 -zxvf helm.tar.gz "linux-amd64/helm"
+  wget -O helm.tar.gz "https://get.helm.sh/helm-v2.17.0-linux-${TARGETARCH}.tar.gz"
+  tar -C /usr/local/bin --strip-components=1 -zxvf helm.tar.gz "linux-${TARGETARCH}/helm"
   mv /usr/local/bin/helm /usr/local/bin/helm2
   rm ./helm.tar.gz
 
   # helm 
-  wget -O helm.tar.gz "https://get.helm.sh/helm-v3.8.2-linux-amd64.tar.gz"
-  tar -C /usr/local/bin --strip-components=1 -zxvf helm.tar.gz "linux-amd64/helm"
+  wget -O helm.tar.gz "https://get.helm.sh/helm-v${HELM_VERSION}-linux-${TARGETARCH}.tar.gz"
+  tar -C /usr/local/bin --strip-components=1 -zxvf helm.tar.gz "linux-${TARGETARCH}/helm"
   rm ./helm.tar.gz
 
   # mutagen
-  wget -O mutagen.tar.gz  "https://github.com/mutagen-io/mutagen/releases/download/v0.11.8/mutagen_linux_amd64_v0.11.8.tar.gz"
-  tar -C /usr/local/bin -zxvf mutagen.tar.gz
-  rm ./mutagen.tar.gz
+  if [ "$TARGETARCH" = amd64 ]; then
+    wget -O mutagen.tar.gz  "https://github.com/mutagen-io/mutagen/releases/download/v0.16.2/mutagen_linux_${TARGETARCH}_v0.11.8.tar.gz"
+    tar -C /usr/local/bin -zxvf mutagen.tar.gz
+    rm ./mutagen.tar.gz
+  fi
 
   # kubeseal
-  wget -O /usr/local/bin/kubeseal https://github.com/bitnami-labs/sealed-secrets/releases/download/v0.12.5/kubeseal-linux-amd64
-  chmod +x /usr/local/bin/kubeseal
+  curl --silent --show-error --fail --location --output kubeseal.tar.gz "https://github.com/bitnami-labs/sealed-secrets/releases/download/v${KUBESEAL_VERSION}/kubeseal-${KUBESEAL_VERSION}-linux-${TARGETARCH}.tar.gz"
+  tar -C /usr/local/bin -zxvf kubeseal.tar.gz kubeseal
+  rm ./kubeseal.tar.gz
+
   addgroup -g 998 docker
   adduser -u 1000 -D ws
   adduser ws docker
 EOF
 
-COPY --from=build "/usr/src/workspace-${WS_VERSION}/ws.phar" /usr/local/bin/ws
+COPY --from=build "/usr/src/workspace/ws.phar" /usr/local/bin/ws
 RUN chmod +x /usr/local/bin/ws && /usr/local/bin/ws
 
 ENTRYPOINT [ "/usr/local/bin/ws" ]
 
-FROM php:7.4-cli-buster as buster
-ARG WS_VERSION=0.2.x
-ARG HELM_VERSION=2.17.0
+FROM php:8.0-cli-buster as buster
+ARG TARGETARCH
+ARG COMPOSE_VERSION
+ARG HELM_VERSION
+ARG KUBESEAL_VERSION
+
+COPY --from=buildx /buildx /usr/libexec/docker/cli-plugins/docker-buildx
 
 RUN <<EOF
   set -o errexit
   set -o nounset
 
-  # docker compose v1
-  curl --silent --show-error --fail --location "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
-  chmod +x /usr/local/bin/docker-compose
-
-  # docker compose v2
-  mkdir -p /usr/libexec/docker/cli-plugins
-  curl --silent --show-error --fail --location -o /usr/libexec/docker/cli-plugins/docker-compose "https://github.com/docker/compose/releases/download/v2.12.1/docker-compose-$(uname -s)-$(uname -m)"
-  chmod +x  /usr/libexec/docker/cli-plugins/docker-compose
-
   echo 'APT::Install-Recommends 0;' >> /etc/apt/apt.conf.d/01norecommends
   echo 'APT::Install-Suggests 0;' >> /etc/apt/apt.conf.d/01norecommends
   apt-get update -qq
+
+  # docker compose v1
+  if [ "$TARGETARCH" = amd64 ]; then
+    curl --silent --show-error --fail --location "https://github.com/docker/compose/releases/download/1.29.2/docker-compose-linux-$(uname -m)" -o /usr/local/bin/docker-compose
+    chmod +x /usr/local/bin/docker-compose
+  else
+    DEBIAN_FRONTEND=noninteractive apt-get -qq -y --no-install-recommends install \
+      python3-bcrypt python3-cryptography python3-pip python3-setuptools python3-dev python3-nacl python3-wheel libffi-dev
+    pip3 install docker-compose
+    apt-get remove -qq -y python3-setuptools python3-dev
+  fi
+
+  # docker compose v2
+  mkdir -p /usr/libexec/docker/cli-plugins
+  curl --silent --show-error --fail --location -o /usr/libexec/docker/cli-plugins/docker-compose "https://github.com/docker/compose/releases/download/v${COMPOSE_VERSION}/docker-compose-$(uname -s)-$(uname -m)"
+  chmod +x  /usr/libexec/docker/cli-plugins/docker-compose
+
   DEBIAN_FRONTEND=noninteractive apt-get -qq -y --no-install-recommends install \
     apt-transport-https ca-certificates curl gnupg
 
   curl --silent --show-error --fail --location https://download.docker.com/linux/debian/gpg | apt-key add -qq - >/dev/null
-  echo "deb [arch=amd64] https://download.docker.com/linux/debian buster stable" > /etc/apt/sources.list.d/docker.list
+  echo "deb [arch=${TARGETARCH}] https://download.docker.com/linux/debian buster stable" > /etc/apt/sources.list.d/docker.list
  
   apt-get update -qq
   DEBIAN_FRONTEND=noninteractive apt-get -s dist-upgrade | grep "^Inst" | \
@@ -120,21 +148,34 @@ RUN <<EOF
       xargs apt-get -qq -y --no-install-recommends install
  
   DEBIAN_FRONTEND=noninteractive apt-get -qq -y --no-install-recommends install \
-   awscli docker-ce-cli git openssh-client jq rsync
+   awscli bash docker-ce-cli git openssh-client jq rsync
    apt-get auto-remove -qq -y
    apt-get clean
    rm -rf /var/lib/apt/lists/*
  
-  curl --silent --show-error --fail --location --output helm.tar.gz "https://get.helm.sh/helm-v${HELM_VERSION}-linux-amd64.tar.gz"
-  tar -C /usr/local/bin --strip-components=1 -zxvf helm.tar.gz "linux-amd64/helm"
+  # helm2
+  curl --silent --show-error --fail --location --output helm.tar.gz "https://get.helm.sh/helm-v2.17.0-linux-${TARGETARCH}.tar.gz"
+  tar -C /usr/local/bin --strip-components=1 -zxvf helm.tar.gz "linux-${TARGETARCH}/helm"
+  mv /usr/local/bin/helm /usr/local/bin/helm2
+  rm ./helm.tar.gz
+
+  # helm 
+  curl --silent --show-error --fail --location --output helm.tar.gz "https://get.helm.sh/helm-v${HELM_VERSION}-linux-${TARGETARCH}.tar.gz"
+  tar -C /usr/local/bin --strip-components=1 -zxvf helm.tar.gz "linux-${TARGETARCH}/helm"
   rm ./helm.tar.gz
 
   # kubeseal
-  curl --silent --show-error --fail --location --output /usr/local/bin/kubeseal https://github.com/bitnami-labs/sealed-secrets/releases/download/v0.12.5/kubeseal-linux-amd64
+  curl --silent --show-error --fail --location --output kubeseal.tar.gz "https://github.com/bitnami-labs/sealed-secrets/releases/download/v${KUBESEAL_VERSION}/kubeseal-${KUBESEAL_VERSION}-linux-${TARGETARCH}.tar.gz"
+  tar -C /usr/local/bin -zxvf kubeseal.tar.gz kubeseal
+  rm ./kubeseal.tar.gz
   chmod +x /usr/local/bin/kubeseal
+
+  groupadd --gid 998 docker
+  useradd --uid 1000 ws
+  usermod --groups docker --append ws
 EOF
 
-COPY --from=build "/usr/src/workspace-${WS_VERSION}/ws.phar" /usr/local/bin/ws
+COPY --from=build "/usr/src/workspace/ws.phar" /usr/local/bin/ws
 RUN chmod +x /usr/local/bin/ws && /usr/local/bin/ws
 
 ENTRYPOINT [ "/usr/local/bin/ws" ]
